requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.21.0,<1.24.0
pillow>=9.5.0
pandas>=1.5.0
requests>=2.26.0
tqdm>=4.65.0
pyyaml>=5.4.0
pyserial>=3.5
scipy>=1.10.0
filterpy>=1.4.5
torchreid==0.2.5
matplotlib>=3.5.0
seaborn>=0.12.0
protobuf<4.21.3
ipython>=8.0.0
psutil>=5.9.0
